requirements.txt
fantraxapi>=1.0.0
MLB-StatsAPI>=1.9.0
pybaseball>=2.2.0
feedparser>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.40.0
pandas>=2.0.0
